Moving actions outside of app
