+++ apps/puzzlify/components/AdventureForm.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { createAdventure } from "@/app/actions/host/create-adventure";
-import { editAdventure } from "@/app/actions/host/edit-adventure";
+import { createAdventure } from "@/actions/host/create-adventure";
+import { editAdventure } from "@/actions/host/edit-adventure";
 import { adventureSchema } from "@/app/schemas/adventure";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useHookFormAction } from "@next-safe-action/adapter-react-hook-form/hooks";
