+++ apps/puzzlify/components/RiddleForm.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { verifyStage } from "@/app/actions/participant/verify-stage";
+import { verifyStage } from "@/actions/participant/verify-stage";
 import { riddleSubmission } from "@/app/schemas/riddle";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useHookFormAction } from "@next-safe-action/adapter-react-hook-form/hooks";
