+++ apps/puzzlify/components/StageForm.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { Adventure, Stage } from "@prisma/client";
-import { createStage } from "@/app/actions/host/create-stage";
+import { createStage } from "@/actions/host/create-stage";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useHookFormAction } from "@next-safe-action/adapter-react-hook-form/hooks";
 import { stageSchema } from "@/app/schemas/stage";
-import { editStage } from "@/app/actions/host/edit-stage";
+import { editStage } from "@/actions/host/edit-stage";
 
 type StageFormProps = {
   adventureId: Adventure["id"];
